Capture/filter Tilt version info
This release captures/filters when the Tilt sends "version" information accompanying a "temperature" of 999 deg F. This prevents both the erroneous logging of incorrect temp/gravity as well as provides the potential to report the version information at a later date.

Additionally, this upgrades the Bluetooth stack to the latest NimBLE version.